+++ src/Jobs/Middleware/SetCorrelation.php
@@ -2,15 +2,16 @@
 
 namespace Butler\Audit\Jobs\Middleware;
 
-use Butler\Audit\Bus\WithCorrelationId;
+use Butler\Audit\Bus\WithCorrelation;
 use Butler\Audit\Facades\Auditor;
 
-class SetCorrelationId
+class SetCorrelation
 {
     public function handle($job, $next)
     {
-        if (in_array(WithCorrelationId::class, class_uses_recursive($job))) {
+        if (in_array(WithCorrelation::class, class_uses_recursive($job))) {
             Auditor::correlationId($job->correlationId);
+            Auditor::correlationTrail($job->correlationTrail);
         }
 
         return $next($job);